+++ projects/silly-datatable/src/lib/shared/models/options-settings.model.ts
@@ -1,0 +1,43 @@
+export class OptionsSettings {
+
+  /**
+   * Css class for options div container.
+   */
+  public containerClass: string;
+
+
+  /**
+   * Css class for checkboxes section div container.
+   */
+  public checkboxSectionClass: string;
+
+
+  /**
+   * Css class for select count of pages section div container.
+   */
+  public selectSectionClass: string;
+
+
+  /**
+   * Css containter class for each checkbox with label.
+   */
+  public checkboxContainerClass: string;
+
+
+  /**
+   * Css class for checkbox
+   */
+  public checkboxClass: string;
+
+
+  /**
+   * Css class for checkbox label
+   */
+  public labelClass: string;
+
+
+  /**
+   * Title (label) text for select of count items per page.
+   */
+  public itemsPerPageTitle: string;
+}