+++ projects/silly-datatable/src/lib/silly-datatable.component.spec.ts
@@ -4,6 +4,9 @@ import { SillyDatatableComponent } from './silly-datatable.component';
 import { ComponentsModule } from './shared/components/components.module';
 import { RequestService } from './shared/services/request.service';
 import { TableParams } from './shared/models/table-params.model';
+import { OptionsService } from './shared/services/options.service';
+import { PipesModule } from './shared/pipes/pipes.module';
+import { OptionsMockService } from './shared/services/options-mock.service';
 
 
 describe('SillyDatatableComponent', () => {
@@ -14,13 +17,18 @@ describe('SillyDatatableComponent', () => {
     TestBed.configureTestingModule({
       imports: [
         ComponentsModule,
+        PipesModule,
       ],
-      declarations: [ SillyDatatableComponent ],
+      declarations: [SillyDatatableComponent],
       providers: [
         RequestService,
+        {
+          provide: OptionsService,
+          useClass: OptionsMockService,
+        },
       ],
     })
-    .compileComponents();
+      .compileComponents();
   }));
 
   beforeEach(() => {