+++ src/app/app.component.spec.ts
@@ -1,8 +1,9 @@
 import { TestBed, async } from '@angular/core/testing';
 import { ReactiveFormsModule } from '@angular/forms';
+import { RouterTestingModule } from '@angular/router/testing';
 
-import { AppComponent } from './app.component';
 import { SillyDatatableModule } from 'silly-datatable';
+import { AppComponent } from './app.component';
 
 
 describe('AppComponent', () => {
@@ -11,6 +12,7 @@ describe('AppComponent', () => {
       imports: [
         SillyDatatableModule,
         ReactiveFormsModule,
+        RouterTestingModule.withRoutes([]),
       ],
       declarations: [
         AppComponent,